+++ lib/core.ps1
@@ -1127,41 +1127,38 @@ function fullpath($path) {
 #       for all communication with api.github.com
 Optimize-SecurityProtocol
 
-# Path gluing has to remaing in these global variables to not fail in case user do not have some environment configured (most likely linux case)
+# TODO: Drop
+$c = get_config 'rootPath'
+if ($c) {
+    Write-UserMessage -Message 'Configuration option ''rootPath'' is deprecated. Configure ''SCOOP'' environment variable instead' -Err
+    if (!$env:SCOOP) { $env:SCOOP = $c }
+}
+
+# Path gluing has to remain in these global variables to not fail in case user do not have some environment configured (most likely linux case)
 # Scoop root directory
-$SCOOP_ROOT_DIRECTORY = $env:SCOOP, (get_config 'rootPath'), "$env:USERPROFILE\scoop" | Where-Object { -not [String]::IsNullOrEmpty($_) } | Select-Object -First 1
-$scoopdir = $SCOOP_ROOT_DIRECTORY
+$SCOOP_ROOT_DIRECTORY = $env:SCOOP, "$env:USERPROFILE\scoop" | Where-Object { -not [String]::IsNullOrEmpty($_) } | Select-Object -First 1
 
 # Scoop global apps directory
-$SCOOP_GLOBAL_ROOT_DIRECTORY = $env:SCOOP_GLOBAL, (get_config 'globalPath'), "$env:ProgramData\scoop" | Where-Object { -not [String]::IsNullOrEmpty($_) } | Select-Object -First 1
-$globaldir = $SCOOP_GLOBAL_ROOT_DIRECTORY
+$SCOOP_GLOBAL_ROOT_DIRECTORY = $env:SCOOP_GLOBAL, "$env:ProgramData\scoop" | Where-Object { -not [String]::IsNullOrEmpty($_) } | Select-Object -First 1
 
 # Scoop cache directory
 # Note: Setting the SCOOP_CACHE environment variable to use a shared directory
 #       is experimental and untested. There may be concurrency issues when
 #       multiple users write and access cached files at the same time.
 #       Use at your own risk.
-$SCOOP_CACHE_DIRECTORY = $env:SCOOP_CACHE, (get_config 'cachePath'), "$SCOOP_ROOT_DIRECTORY\cache" | Where-Object { -not [String]::IsNullOrEmpty($_) } | Select-Object -First 1
-$cachedir = $SCOOP_CACHE_DIRECTORY
-
-# Scoop config file migration
-$configHome = $env:XDG_CONFIG_HOME, "$env:USERPROFILE\.config" | Select-Object -First 1
-$SCOOP_CONFIGURATION_FILE = Join-Path $configHome 'scoop\config.json'
-$configFile = $SCOOP_CONFIGURATION_FILE
-$oldConfigPath = Join-Path $env:USERPROFILE '.scoop'
-
-if ((Test-Path $oldConfigPath) -and !(Test-Path $SCOOP_CONFIGURATION_FILE)) {
-    Split-Path $SCOOP_CONFIGURATION_FILE -Parent | ensure | Out-Null
-    Move-Item $oldConfigPath $SCOOP_CONFIGURATION_FILE
-    Write-UserMessage -Warning -Message @(
-        "Scoop configuration has been migrated from '~/.scoop'"
-        "to '$SCOOP_CONFIGURATION_FILE'"
-    )
-}
+$SCOOP_CACHE_DIRECTORY = $env:SCOOP_CACHE, "$SCOOP_ROOT_DIRECTORY\cache" | Where-Object { -not [String]::IsNullOrEmpty($_) } | Select-Object -First 1
 
 # Load Scoop config
+$configHome = $env:XDG_CONFIG_HOME, "$env:USERPROFILE\.config" | Where-Object { -not [String]::IsNullOrEmpty($_) } | Select-Object -First 1
+$SCOOP_CONFIGURATION_FILE = Join-Path $configHome 'scoop\config.json'
 $SCOOP_CONFIGURATION = load_cfg $SCOOP_CONFIGURATION_FILE
+
+# TODO: Remove deprecated variables
+$scoopdir = $SCOOP_ROOT_DIRECTORY
+$globaldir = $SCOOP_GLOBAL_ROOT_DIRECTORY
+$cachedir = $SCOOP_CACHE_DIRECTORY
 $scoopConfig = $SCOOP_CONFIGURATION
+$configFile = $SCOOP_CONFIGURATION_FILE
 
 # Setup proxy globally
 setup_proxy